requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
tensorflow==2.16.1
numpy==1.26.4
pillow==10.3.0
opencv-python-headless==4.10.0.84
python-multipart==0.0.9
